Update dev-setup script with brew installs

+++ dev-setup.sh
@@ -1,23 +1,72 @@
 #!/usr/bin/env bash
 GIT_CONFIG="$HOME/.gitconfig"
-
-echo "NAME for git? "
-read NAME
-git config --global user.name "$NAME"
-
-echo "EMAIL for git? "
-read EMAIL
-git config --global user.email "$EMAIL"
-
-sed "/$NAME/ s/$/ #gitignore/" "${GIT_CONFIG}" > "${GIT_CONFIG}_tmp"
-sed "/$EMAIL/ s/$/ #gitignore/" "${GIT_CONFIG}_tmp" > "${GIT_CONFIG}_tmp2"
-sed '/\[user\]/ s/$/ #gitignore/' "${GIT_CONFIG}_tmp2" > "${GIT_CONFIG}"
-
-rm "${GIT_CONFIG}_tmp"
-rm "${GIT_CONFIG}_tmp2"
-
-# Install Homebrew
-/usr/bin/ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
+CONFIRM_ALL=${1}
+GIT_USERNAME=$(git config user.name)
+GIT_EMAIL=$(git config user.email)
 
 
+function caskInstall {
+    if [[ "$CONFIRM_ALL" != "y" ]]; then
+        echo "Install $1 (y|n)?"
+        read in
+
+        if [[ "$in" != "y" ]]; then
+            return
+        fi;
+    fi;
+    echo "Installing $1..."
+    brew cask install $1
+}
+
+
+function brewInstall {
+    if [[ "$CONFIRM_ALL" != "y" ]]; then
+        echo "Install $1 (y|n)?"
+        read in
+
+        if [[ "$in" != "y" ]]; then
+            return
+        fi;
+    fi;
+    brew install $1
+}
+
+
+if [[ "$GIT_USERNAME" == "" ]] && [[ "$GIT_EMAIL" == "" ]]; then
+
+    echo "NAME for git? "
+    read NAME
+    git config --global user.name "$NAME"
+
+    echo "EMAIL for git? "
+    read EMAIL
+    git config --global user.email "$EMAIL"
+
+    sed "/$NAME/ s/$/ #gitignore/" "${GIT_CONFIG}" > "${GIT_CONFIG}_tmp"
+    sed "/$EMAIL/ s/$/ #gitignore/" "${GIT_CONFIG}_tmp" > "${GIT_CONFIG}_tmp2"
+    sed '/\[user\]/ s/$/ #gitignore/' "${GIT_CONFIG}_tmp2" > "${GIT_CONFIG}"
+    rm "${GIT_CONFIG}_tmp"
+    rm "${GIT_CONFIG}_tmp2"
+fi;
+
+if [[ "$CONFIRM_ALL" == "y" ]]; then
+    echo "skipping install confirmations..."
+fi;
+
+
+which brew
+if [[ "$?" -eq "1" ]]; then
+    # Install Homebrew
+    echo "Installing Homebrew..."
+    /usr/bin/ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
+fi;
+
+brew update
+caskInstall google-chrome
+caskInstall iterm2
+caskInstall visual-studio-code
+caskInstall slack
+caskInstall spotify
+caskInstall docker
+brewInstall rg
